Prerobenie na Double

+++ src/main/java/org/example/Controller.java
@@ -116,7 +116,7 @@ public class Controller {
         }
 
         var userId = Auth.INSTANCE.getPrincipal().getId();
-                BigDecimal currentBalance = userDAO.getBalance(userId);
+                Double currentBalance = userDAO.getBalance(userId);
                 if (currentBalance != null) {
                     Auth.INSTANCE.getPrincipal().setBalance(currentBalance); // Aktualizácia balansu v Auth
                     loginoruser.setText(Auth.INSTANCE.getPrincipal().getUsername() + "\nZostatok: " + currentBalance);
