+++ src/main/java/org/example/statistics/Statistics.java
@@ -2,7 +2,7 @@ package org.example.statistics;
 import lombok.Data;
 @Data
 public class Statistics {
-    private long UserId;
+    private Integer UserId;
     private int totalBets;
     private double totalStakes;
     private double totalWinnings;
